+++ emabot/backtest.py
@@ -19,12 +19,26 @@ def main() -> None:
     parser = argparse.ArgumentParser()
     parser.add_argument('--csv-file',
         help='Path to OHLC CSV file', dest='csv_file', required=True)
+    parser.add_argument('--resample',
+        help='Set the OHLC resample size (default:1D)', dest='resample', required=False,
+        default='1D')
+    parser.add_argument('--ema-a',
+        help='Set the emaA paramater (default:1)', dest='ema_a', required=False, default=1,
+        type=int)
+    parser.add_argument('--ema-b',
+        help='Set the emaB paramater (default:2)', dest='ema_b', required=False, default=2,
+        type=int)
     parser.add_argument('--debug',
         help='Enable debug output', dest='debug', action='store_true', required=False)
     parser.add_argument('--c2c',
         help='Coin-to-coin (higher precision)', dest='c2c', action='store_true', required=False)
     args = parser.parse_args()
-    stats = backtest(emaA=1, emaB=2, csv_file=args.csv_file, c2c=args.c2c, debug=args.debug)
+    print('ARGS: emaA={} emaB={} resample={} c2c={}'.format(
+        args.ema_a, args.ema_b, args.resample, args.c2c))
+    stats = backtest(
+        emaA=args.ema_a, emaB=args.ema_b, resample=args.resample,
+        csv_file=args.csv_file,
+        c2c=args.c2c, debug=args.debug)
     fee_total = sum([sum(i) for i in stats.per_day['fee'].values()])
     percent_total = sum([sum(i) for i in stats.per_day['percent'].values()])
     net_profit_total = sum([sum(i) for i in stats.per_day['net_profit'].values()])